+++ tests/e2e/registry/test_registry_pod_respin.py
@@ -1,7 +1,6 @@
-import shlex
 import pytest
 import logging
-from subprocess import Popen, PIPE
+
 
 from ocs_ci.ocs.constants import OPENSHIFT_IMAGE_REGISTRY_NAMESPACE
 from ocs_ci.ocs.exceptions import UnexpectedBehaviour
@@ -11,35 +10,12 @@ from ocs_ci.ocs.registry import (
     check_image_exists_in_registry
 )
 from ocs_ci.framework.testlib import E2ETest, workloads
-from ocs_ci.utility.svt import (
-    svt_project_clone, svt_create_venv_setup,
-    svt_cluster_loader, svt_cleanup
-)
+from ocs_ci.utility.svt import svt_setup, svt_cleanup
 from tests import disruption_helpers
 from tests.sanity_helpers import Sanity
 
 log = logging.getLogger(__name__)
 IMAGE_URL = 'docker.io/library/busybox'
-
-
-def svt_setup(iterations):
-    """
-    Creates and run svt workload
-    """
-
-    svt_project_clone()
-    svt_create_venv_setup()
-    svt_cluster_loader()
-    cmd = (
-        "/bin/sh -c 'source /tmp/venv/bin/activate && "
-        f"python /tmp/svt/openshift_performance/ose3_perf/scripts/build_test.py -z -a -n {iterations}'"
-    )
-    log.info(f"Running command {cmd}")
-    build = Popen(shlex.split(cmd), stdout=PIPE, stderr=PIPE)
-    stdout, stderr = build.communicate()
-    out = stderr.split('\n'.encode())[-10].split()[-1]
-    log.info(stderr)
-    assert (int(out) == 0), 'Failed builds found'
 
 
 @workloads
